Update kmain to use the more hip u64 instead of size_t

+++ arch/i686-elf/kmain.cpp
@@ -51,7 +51,7 @@ extern "C"
     // make sure it's at least 1MB
     if(mem_total < MIN_MEM)
     {
-      TERMINST().printf("error: I need at least % kb of memory to run", static_cast<size_t>(MIN_MEM));
+      TERMINST().printf("error: I need at least % kb of memory to run", static_cast<u64>(MIN_MEM));
       halt();
     }
